coin/requirements.txt
RPi.GPIO
zbarlight
Pillow